[DX] Move MongoDB Document normalizers
