+++ src/Pim/Bundle/CatalogBundle/MongoDB/Normalizer/Document/ProductNormalizer.php
@@ -1,8 +1,13 @@
 <?php
 
-namespace Pim\Bundle\TransformBundle\Normalizer\MongoDB;
+namespace Pim\Bundle\CatalogBundle\MongoDB\Normalizer\Document;
 
 use Akeneo\Bundle\StorageUtilsBundle\MongoDB\MongoObjectsFactory;
+use Pim\Bundle\TransformBundle\Normalizer\MongoDB\Association;
+use Pim\Bundle\TransformBundle\Normalizer\MongoDB\Category;
+use Pim\Bundle\TransformBundle\Normalizer\MongoDB\Collection;
+use Pim\Bundle\TransformBundle\Normalizer\MongoDB\Group;
+use Pim\Bundle\TransformBundle\Normalizer\MongoDB\ProductValue;
 use Pim\Component\Catalog\Model\ProductInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 use Symfony\Component\Serializer\SerializerAwareInterface;
