+++ src/Pim/Bundle/CatalogBundle/MongoDB/Normalizer/Document/ProductPriceNormalizer.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Pim\Bundle\TransformBundle\Normalizer\MongoDB;
+namespace Pim\Bundle\CatalogBundle\MongoDB\Normalizer\Document;
 
 use Akeneo\Bundle\StorageUtilsBundle\MongoDB\MongoObjectsFactory;
+use Pim\Bundle\CatalogBundle\MongoDB\Normalizer\Document\ProductNormalizer;
 use Pim\Component\Catalog\Model\ProductPriceInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
