+++ src/Pim/Bundle/ReferenceDataBundle/Resources/config/storage_driver/doctrine/mongodb-odm.yml
@@ -2,7 +2,7 @@ parameters:
     pim_reference_data.doctrine.query.filter.reference_data.class: Pim\Bundle\ReferenceDataBundle\Doctrine\MongoDB\Filter\ReferenceDataFilter
     pim_reference_data.doctrine.query.sorter.reference_data.class: Pim\Bundle\ReferenceDataBundle\Doctrine\MongoDB\Sorter\ReferenceDataSorter
     pim_reference_data.mongodb.normalizer.reference_data.class:    Pim\Bundle\ReferenceDataBundle\MongoDB\Normalizer\NormalizedData\ReferenceDataNormalizer
-    pim_reference_data.normalizer.mongodb.reference_data.class:    Pim\Bundle\ReferenceDataBundle\Normalizer\MongoDB\ReferenceDataNormalizer
+    pim_reference_data.normalizer.mongodb.reference_data.class:    Pim\Bundle\ReferenceDataBundle\MongoDB\Normalizer\Document\ReferenceDataNormalizer
 
 services:
     pim_reference_data.mongodb.normalizer.reference_data:
